NavTreeItem/NavTreeFolder implementation
Reworked how items show up in the NavTree.

OLD idea was to have items show as 'CatalogItems'.

NEW idea is to have separate class for items showing in NavTree. For
now, these are NavTreeItemBase and NavTreeFolder.

NavTreeItemBase just defines some common properties. Most importantly
are ChildContainers : a list of child NavTreeItems in the current; and
ChildItems : a list of items in the current. Think of these like
Directories and Files, respectively.

NavTreeFolder implements these properties. DirInfo / FileInfo provides
most of the logic. It also allows a regex search pattern so that only
ChildItems matching certain patterns are returned. Good for matching
only *jpgs, for instance.

The UI is updated to reflect some of these changes.