Catalog Class; Related Tests
This checkin adds a class 'Catalog'. This class is a controller for some
of the simpler classes made so far. It contains a TagRegister, and
Dictionary of DisplayItemProperties. The idea is to use this as a
central location for mapping files to properties.

Design idea is for Information about the file added by the user to be
tracked through Catalog and then through DisplayItemProperties, with
TagRegister assisting in lookup. (TagRegister should probably be
private). Information about items during runtime (like image size,
width/height, basically anything we read from disk) will be stored in
the class DisplayItem (probably will end up renaming this to
DisplayItemRunTimeInfo.

First tests used to excercise these classes added as well