Add-NavTree-Folder Commands and DataValidation
Added quite a bit of properties to ImageCatalogViewModel in order to
allow addition of Folders through UI. Validation using IDataErrorInfo
exists as well. The commands stilll need to be re-implemented as
RelayCommands.

Wired the new view model up to the view.

Added NavTree folder validation on adding new directories. We now check
if we can iterate folders in the dir. If we can't, it is assumed
unauthorized, and we won't add it to the UI.