RelayCommand added and used
Added the 'RelayCommand' pattern from Josh Smith MVVM example.

Wired up the RelayCommand to the previous add new folder command. This
cleaned up the constructor a bit and removed the need to explicitly
check whether the button is valid or not in XAML. Button switch from
valid/invalid seems a little slower now, but not a big thing.