Catalog Save/Load to Json
This change now saves the Catalog to disk.

The object is saved using the Json serialize/deserialize methods. Quite
a few properties had to be marked as 'JsonProperty' or 'JsonConstructor'
to get this to work. Also, attempt was made to limit the amount saved to
disk (don't need to save every file, just the dirs, etc).

Tests added to verify.

Application moved from 'UserControl' to 'MainWindow'.