DisplayItem async runtime properties, DisplayItem View
This change does the following :

DisplayItem Properties are now divided between 'RunTime' and 'User'
properties. The first is info we can find at runtime, by examining the
file. It is not stored because we don't want to waste storage saving it,
it may change over time. The second is info we store/save in catalog.

DisplayItemRunTime properties are for the most part loaded
asynchronously. Seems to work. The problem now is that all the Image
objects take up a lot of memory, and we quicky see memory balloon as
images are read from disk.

DIsplayProperty View added to UI.