DisplayItem UI and Tag Modification, ViewModel Craziness.
Added a number of new ViewModels, we should bind to these from now on,
NOT properties/methods on the Model.

Renamed 'DisplayItemProperties' to 'DisplayItemUserProperties'.
Properties are now divided between user/runtime properties. The first
are saved to disk, the second accessed at runtime.

Made TagRegister Internal. Maintaining the mapping between tags and
files is complicated, Catalog does it correctly, but exposing the
methods elsewhere is error prone especially if open to UI. Thus,
disabling.

Needed to add various references to the catalog to a number of items.