Tag Navigation
Wow did things suddenly get hacky. Idea is to add another nav item under
the folders that would let users search by tag. This ran into a few
problems, and exposed some more.

MVVM Issues:
- The tree view item has no concept of selected item. Add a 'behavior'
class from SO that fixes this.

- The tree view item absolutely does not support multi selection. Gave
up on this altogether.

- Syncing the Display between 2 items (tags and folders). This works,
but required a lot of trickery and non-mvvm stuff

- SelectionMode=Extended doesnt reliably trigger SelectedItem binding.
Hacked through this with an 'IsSelected' property on NavFolderViewModel.

- Tags updated in a DisplayItemViewModel have no way of notifying
ImageCatalogViewModel of property changed.

The feeling gained here is that understanding of MVVM is not entirely
baked, and question if MVVM can do everything desired of it. Working
through these issues required events, weird binding tricks, and just
generally seemed more painful than it should be.

Additional problem found in image loading :
- At least when debugging, loading up all the image objects in
RunTimeProperties is error prone. OutOfMemoryExceptions and Files Locked
are the big ones.

Putting this on hold, because the image loading needs a re-haul.